startable_stoppable_game: Make Decision class to interact with the Behavior Class
A behavior is what makes someone respond to something. This can be modeled with
a condition -> consequence paradigm. When a condition is met, the piece makes a
decision about what to do, i.e. what will be the consequence for the condition
being met.

It doesn't make a lot of sense at the moment. :/